requirements.txt
-e file:///Users/tahmidahamed/Dev/amazon-agent
annotated-types==0.7.0
anyio==4.8.0
attrs==25.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
griffe==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.9.0
openai==1.66.2
openai-agents==0.0.3
outcome==1.3.0.post0
packaging==24.2
pip==25.0.1
pluggy==1.5.0
pydantic==2.10.6
pydantic-core==2.27.2
pysocks==1.7.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
python-dotenv==1.0.1
requests==2.32.3
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
types-requests==2.32.0.20250306
typing-extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
